requirements.txt
# 核心依赖
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# 前端
streamlit==1.28.0

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0

# LightRAG
lightrag>=0.1.0

# 模型相关
openai==1.3.0

# Agent 框架
langgraph>=0.0.20
langchain>=0.1.0

# 爬虫
playwright==1.40.0
scrapy==2.11.0

# 配置管理
pyyaml==6.0.1
python-dotenv==1.0.0

# 工具库
requests==2.31.0
aiohttp==3.9.1

# 日志和监控
python-json-logger==2.0.7

